update to newest Vulcan syntax
