+++ lib/client/interface.js
@@ -69,58 +69,58 @@ function serializeFormData(obj, formDataObj = {}, namespace) {
 }
 
 // DEPRECATED with apollo 2
-export const fileUploadMiddleware = {
-  applyMiddleware({ request, options }, next) {
-    if (isUpload(request)) {
-      const body = new FormData();
-      const data = [];
-      const printed = printRequest(request);
-      data.push({
-        operationName: printed.operationName || undefined,
-        debugName: printed.debugName || undefined,
-        query: printed.query,
-        variables: request.variables || {}
-      });
-      const serialised = serializeFormData({ data });
-      Object.entries(serialised).forEach(([name, value]) => {
-        if (typeof value !== typeof undefined) {
-          body.set(name, value);
-        }
-      });
-      options.headers = options.headers || new Headers();
-      options.headers.Accept = '*/*';
-      options.headers['Content-Type'] = undefined;
-      options.body = body;
-    }
-    next();
-  },
-  applyBatchMiddleware({ requests, options }, next) {
-    if (isUpload(requests)) {
-      const body = new FormData();
-      const data = [];
-      requests.forEach((request, i) => {
-        const printed = printRequest(request);
-        data.push({
-          operationName: printed.operationName || undefined,
-          debugName: printed.debugName || undefined,
-          query: printed.query,
-          variables: request.variables || {}
-        });
-      });
-      const serialised = serializeFormData({ data });
-      Object.entries(serialised).forEach(([name, value]) => {
-        if (typeof value !== typeof undefined) {
-          body.set(name, value);
-        }
-      });
-      options.headers = options.headers || new Headers();
-      options.headers.Accept = '*/*';
-      options.headers['Content-Type'] = undefined;
-      options.body = body;
-    }
-    next();
-  }
-};
+//export const fileUploadMiddleware = {
+//  applyMiddleware({ request, options }, next) {
+//    if (isUpload(request)) {
+//      const body = new FormData();
+//      const data = [];
+//      const printed = printRequest(request);
+//      data.push({
+//        operationName: printed.operationName || undefined,
+//        debugName: printed.debugName || undefined,
+//        query: printed.query,
+//        variables: request.variables || {}
+//      });
+//      const serialised = serializeFormData({ data });
+//      Object.entries(serialised).forEach(([name, value]) => {
+//        if (typeof value !== typeof undefined) {
+//          body.set(name, value);
+//        }
+//      });
+//      options.headers = options.headers || new Headers();
+//      options.headers.Accept = '*/*';
+//      options.headers['Content-Type'] = undefined;
+//      options.body = body;
+//    }
+//    next();
+//  },
+//  applyBatchMiddleware({ requests, options }, next) {
+//    if (isUpload(requests)) {
+//      const body = new FormData();
+//      const data = [];
+//      requests.forEach((request, i) => {
+//        const printed = printRequest(request);
+//        data.push({
+//          operationName: printed.operationName || undefined,
+//          debugName: printed.debugName || undefined,
+//          query: printed.query,
+//          variables: request.variables || {}
+//        });
+//      });
+//      const serialised = serializeFormData({ data });
+//      Object.entries(serialised).forEach(([name, value]) => {
+//        if (typeof value !== typeof undefined) {
+//          body.set(name, value);
+//        }
+//      });
+//      options.headers = options.headers || new Headers();
+//      options.headers.Accept = '*/*';
+//      options.headers['Content-Type'] = undefined;
+//      options.body = body;
+//    }
+//    next();
+//  }
+//};
 
 export function createUploadNetworkInterface(opts) {
   const batchedInterface = createBatchingNetworkInterface(opts);
