+++ lib/server/fieldResolver.js
@@ -1,9 +1,9 @@
-import createResolverMultiple from './createResolverMultiple';
-import createResolverSingle from './createResolverSingle';
 import defaultResolveId from '../modules/defaultResolveId';
+import castArray from 'lodash/castArray';
+import compact from 'lodash/compact';
 
 /**
- * Creates field's resolver.
+ * Creates field's resolver to retrieve the file.
  *
  * @param {Object} options Options
  * @param {String} options.fieldName
@@ -28,4 +28,34 @@ export default function createResolver(options) {
   return multiple
     ? createResolverMultiple(fieldName, FSCollection, resolveId)
     : createResolverSingle(fieldName, FSCollection, resolveId);
+}
+
+function createResolverMultiple(
+  fieldName,
+  FSCollection,
+  resolveId,
+) {
+  return async document => {
+    const fileIds = compact(castArray(document[fieldName])).map(resolveId);
+    if (fileIds.length === 0) return [];
+
+    const files = await FSCollection.loader.loadMany(fileIds);
+    return compact(files);
+  };
+}
+
+function createResolverSingle(
+  fieldName,
+  FSCollection,
+  resolveId,
+) {
+  const resolverMultiple = createResolverMultiple(
+    fieldName,
+    FSCollection,
+    resolveId,
+  );
+  return async document => {
+    const files = await resolverMultiple(document);
+    return files[0] || null;
+  }
 }
