+++ lib/server/generateFieldSchema.js
@@ -1,8 +1,8 @@
 import merge from 'lodash/merge';
 
 import generateFieldSchemaBase from '../modules/generateFieldSchemaBase';
-import createHandlers from './createHandlers';
-import createResolver from './createResolver';
+import fieldUploadHandlers from './fieldUploadHandlers';
+import fieldResolver from './fieldResolver';
 
 /**
  * Server's function to generate a field schema. What this basically does is to
@@ -23,9 +23,9 @@ export default (options = {}) => {
   return generateFieldSchemaBase({
     ...options,
     fieldSchema: merge({}, fieldSchema, {
-      ...createHandlers(options),
+      ...fieldUploadHandlers(options), // onCreate / onUpdate callbacks to upload the file
       resolveAs: {
-        resolver: createResolver(options),
+        resolver: fieldResolver(options),
       },
     }),
   });
