+++ lib/server/main.js
@@ -1,4 +1,3 @@
-import './uploadMiddleware';
 import './addGraphQLSchemaAndResolvers';
 
 export * from '../modules';
